# Conflicts:
#	Tools/autotest/arduplane.py
#	libraries/SITL/SIM_Aircraft.h

+++ libraries/AP_Compass/AP_Compass_LIS3MDL.cpp
@@ -113,7 +113,7 @@ bool AP_Compass_LIS3MDL::init()
     }
     set_dev_id(compass_instance, dev->get_bus_id());
 
-    printf("Found a LIS3MDL on 0x%x as compass %u\n", dev->get_bus_id(), compass_instance);
+    printf("Found a LIS3MDL on 0x%x as compass %u\n", unsigned(dev->get_bus_id()), compass_instance);
     
     set_rotation(compass_instance, rotation);
